# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Connect | Server",
    "version": "2.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/ambulances/documents": {
      "patch": {
        "operationId": "updateAmbulanceDocumentsInBulk",
        "summary": "Update multiple ambulance documents",
        "tags": [
          "Ambulance"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "documents": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "cuid",
                          "pattern": "^[cC][^\\s-]{8,}$"
                        },
                        "name": {
                          "type": "string"
                        },
                        "downloadUrl": {
                          "type": "string",
                          "format": "uri"
                        },
                        "fileExtension": {
                          "type": "string"
                        },
                        "fileSize": {
                          "type": "number"
                        },
                        "ambulanceId": {
                          "type": "string"
                        },
                        "expiresAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "downloadUrl",
                        "fileExtension",
                        "fileSize",
                        "ambulanceId",
                        "expiresAt"
                      ]
                    }
                  },
                  "baseId": {
                    "type": "string"
                  }
                },
                "required": [
                  "documents",
                  "baseId"
                ]
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updatedCount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "updatedCount"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/ambulances/{id}/switch-status": {
      "patch": {
        "operationId": "switchAmbulanceStatus",
        "summary": "Switch an ambulance status",
        "tags": [
          "Ambulance"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "QAP",
                      "OCP",
                      "EVT",
                      "J4",
                      "J5",
                      "FA",
                      "MN"
                    ]
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/bases/{id}": {
      "put": {
        "operationId": "updateBase",
        "summary": "Update base",
        "tags": [
          "Base"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "document": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getBase",
        "summary": "Get base",
        "tags": [
          "Base"
        ],
        "description": "Get base",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "document": {
                          "nullable": true,
                          "type": "string"
                        },
                        "latitude": {
                          "nullable": true,
                          "type": "number"
                        },
                        "longitude": {
                          "nullable": true,
                          "type": "number"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "phones": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "cuid",
                                "pattern": "^[cC][^\\s-]{8,}$"
                              },
                              "number": {
                                "type": "string"
                              },
                              "isWhatsapp": {
                                "type": "boolean"
                              },
                              "name": {
                                "nullable": true,
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "id",
                              "number",
                              "isWhatsapp",
                              "name",
                              "createdAt"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "unit": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "cuid",
                              "pattern": "^[cC][^\\s-]{8,}$"
                            },
                            "name": {
                              "type": "string"
                            },
                            "company": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "format": "cuid",
                                  "pattern": "^[cC][^\\s-]{8,}$"
                                },
                                "name": {
                                  "type": "string"
                                },
                                "document": {
                                  "type": "string"
                                },
                                "companyGroup": {
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "type": "string",
                                      "format": "cuid",
                                      "pattern": "^[cC][^\\s-]{8,}$"
                                    },
                                    "name": {
                                      "type": "string"
                                    },
                                    "document": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "name",
                                    "document"
                                  ],
                                  "additionalProperties": false
                                }
                              },
                              "required": [
                                "id",
                                "name",
                                "document",
                                "companyGroup"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "company"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "document",
                        "latitude",
                        "longitude",
                        "createdAt",
                        "updatedAt",
                        "phones",
                        "unit"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteBase",
        "summary": "Delete base",
        "tags": [
          "Base"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}